% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-nc.R
\name{src_nc}
\alias{src_nc}
\title{NetCDF virtual tables}
\usage{
src_nc(ncsource)
}
\arguments{
\item{ncsource}{}
}
\value{
a 'src_nc', a virtual-database for a NetCDF source
}
\description{
NetCDF virtual tables
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
src_nc(f)
\dontrun{
nc <- src_nc("http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdQSwind3day")
nc
}
}
