% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchelper.R
\name{[.nchelper}
\alias{[.nchelper}
\title{Extract}
\usage{
\method{[}{nchelper}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{nchelper} object}

\item{...}{index arguments, see \code{base::Extract}}

\item{drop}{remove degenerate singleton dimensions, \code{TRUE} by default}
}
\description{
Extract
}
