% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchelper.R
\name{nchelper}
\alias{nchelper}
\title{nchelper}
\usage{
nchelper(file, varname = NULL)
}
\arguments{
\item{file}{NetCDF file}

\item{varname}{NetCDF variable to access}
}
\value{
nchelper object
}
\description{
Helper functions to treat a NetCDF variable as an R array
}
\details{
Create an object to represent a NetCDF variable, primarily to access slices from it with R's Extract syntax for arrays.
}
\examples{
f <- system.file("extdata", "avhrr-only-v2.20180126.nc", package = "nchelper")
a <- nchelper(f, "sst")
plot(a[,360,,] * 0.01, ylim = c(20, 32), type = "l")
}
